separating websocket_server.py into multiple files
