+++ search_term.md
@@ -1,1 +1,1 @@
-Recurrent Bladder Cancer
+Metastatic Bladder Cancer
